requirements/base.txt
pytz
python-slugify
Pillow
argon2-cffi
redis
hiredis

# Django
# ------------------------------------------------------------------------------
django
django-environ
django-model-utils
django-allauth
django-crispy-forms
crispy-bootstrap5
django-redis
# Django REST Framework
djangorestframework
django-cors-headers
drf-writable-nested>=0.7.0  #
# DRF-spectacular for api documentation
drf-spectacular
